+++ src/OiLaravelDocumentationServiceProvider.php
@@ -7,6 +7,7 @@ use OiLab\LaravelDocumentation\Console\Commands\AddDocPage;
 use OiLab\LaravelDocumentation\Console\Commands\GenDocIndex;
 use OiLab\LaravelDocumentation\Console\Commands\GenDocNav;
 use OiLab\LaravelDocumentation\Console\Commands\ImportPackageDocs;
+use OiLab\LaravelDocumentation\Console\Commands\InstallAiSkill;
 use OiLab\LaravelDocumentation\Console\Commands\InstallDocumentation;
 use OiLab\LaravelDocumentation\Services\DocumentationService;
 
@@ -45,6 +46,7 @@ class OiLaravelDocumentationServiceProvider extends ServiceProvider
                 InstallDocumentation::class,
                 AddDocPage::class,
                 ImportPackageDocs::class,
+                InstallAiSkill::class,
             ]);
         }